fix(api): create securityGate.js to store all security check functions
