+++ server/api/users.js
@@ -2,8 +2,7 @@ const router = require('express').Router()
 const {User} = require('../db/models')
 module.exports = router
 
-const isAdmin = (req, res, next) =>
-  req.user.isAdmin ? next() : res.send('None shall pass')
+const {isAdmin, isYourself} = require('./securityGate')
 
 router.get('/', isAdmin, async (req, res, next) => {
   try {
@@ -15,16 +14,6 @@ router.get('/', isAdmin, async (req, res, next) => {
     next(err)
   }
 })
-
-const isYourself = (req, res, next) => {
-  console.log('params', req.params.id)
-  console.log('user', req.user.id)
-  if (+req.params.id === req.user.id) {
-    next()
-  } else {
-    res.send("you don't have access to this user")
-  }
-}
 
 router.get('/:id', isYourself, async (req, res, next) => {
   try {
